pyproject.toml
[tool.poetry]
name = "ibmodules"
version = "0.1.0"
description = ""
authors = ["Iron Bank Pipelines and Operations <ironbank@dsop.io>"]

packages = [
    { include = "ironbank" }
]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.135"
gitpython = "^3.1.27"
jsonschema = "^4.6.0"
openpyxl = "^3.0.10"
pandas = "^1.4.2"
python-gitlab = "^3.5.0"
requests = "^2.28.0"
pyyaml = "^6.0"
yq = "^3.2.2"
isort = "^5.12.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
ipython = "^8.11.0"
mypy = "^1.0.0"
jinja2 = "^3.1.2"
pylama = "^8.4.1"
pylint = "^2.16.1"
pysocks = "^1.7.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
selenium = "^4.8.0"
types-pyyaml = "^6.0.12.4"
types-requests = "^2.28.11.8"
autoflake = "^2.1.1"
anybadge = "^1.14.0"
docformatter = "^1.7.2"

[build-system]
requires = ["poetry-core==1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = "True"
check_untyped_defs = "True"
show_error_codes = "True"
warn_unused_ignores = "True"
no_implicit_optional = "True"
warn_return_any = "True"
ignore_missing_imports = "True"
exclude = ['test_',]

[tool.radon]
exclude = "test_*.py"
cc_min = "D"
show_complexity = "True"

[tool.isort]
line_length = 80